Make pipeWrte and pipeRead noexcept
After the previous device maps commits, `pipeWrite` might throw. In
this case, if we increment active calls before `pipeWrite` on the
caller, that active call won't be decremented properly when `pipeWrite`
throws. As a result, `shutdown` can silently timeout. I noticed this
as some tests take more than 60s to finish.

This commit extract the tensor device checking logic out of pipeWrite,
and make sure the error is thrown before the active call count is
incremented.

ghstack-source-id: 24ee248915a4274b5f736d4cbf522d06d406dce1
Pull Request resolved: https://github.com/pytorch/pytorch/pull/45783